Moved Shortcuts on a global file
