--- src/lib/ui/PsView/shortcut.ts
@@ -1,5 +1,0 @@
-import { Keyboard } from "@raycast/api";
-
-export const Shortcut = {
-  ToggleQuickLook: Keyboard.Shortcut.Common.ToggleQuickLook,
-} as const;
